requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
requests>=2.30.0